requirements.txt
greenlet==3.3.2
iniconfig==2.3.0
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.20.0
pytest==9.0.3
PyYAML==6.0.3
typing_extensions==4.15.0
